Допиливание сервиса, добавление парсера сайта inshaker

+++ internal/domain/models/models.go
@@ -1,6 +1,6 @@
 package models
 
 type Pagination struct {
-	Page         int
-	ItemsPerPage int
+	Page         int64
+	ItemsPerPage int64
 }
